fix: markdown preview only rendered when active
