+++ backend/theme/forms.py
@@ -1,6 +1,5 @@
 from django import forms
 from .models import ThemeConfig
-from experiment.forms import MarkdownPreviewTextInput
 
 
 class ThemeConfigForm(forms.ModelForm):
@@ -21,4 +20,3 @@ class ThemeConfigForm(forms.ModelForm):
         if commit:
             instance.save()
         return instance
-
